// src/app/modules/mantenimiento/models/V-Producto-usuario.model.ts
export class VProductoUsuario
 {
    idProducto: number = 0;
    Categoria: string = "";
    idCategoria: number = 0;
    subcategoria: string  = "";
    idSubcategoria: number  = 0;
    nombre: string  = "";
    descripcion: string  = "";
    imagen: string  = "";
    precio: number  = 0;
    estado: boolean  = false;

}